Huge refactoring to improve #2
Configurations are now handled in a more decent way.
Using OptionParser for options.
All validation errors inside the configuration class.
Delivery decision/loading/configuration is now also within the
configuration class so initializers died.
Discovered that the sender of the emails is a completely separate
property that probably needs to be put somewhere.
Renamed the bulk processor to be a worker. Could facilitate if we ever
move to a sidekiq or resque or similar solution.
Testing those new classes should be a lot easier now.