One more big refactoring and test adding for #2
This is now pretty close to as good as I want it to be.
Need to add tests to the configuration class and then move on to other
things.
Attendee is the most screwed class at this point. Need to rearchitect
that whole thing.
Got rid of the extension again. When things are nicely organized,
requiring individual files is easy and painless.