Big commit to clean up some things
Close #4 by creating the data_folder idea. This folder should have 3
files: model.svg, data.csv and email.md.erb.
Model.svg is the SVG template.
Data.csv is the CSV with all attendee's data.
Email.md.erb is the body template to be interpreted in ERB and then
transformed to HTML by a markdown parser to generate the email subject,
the email body in plain text and the email body in html.
Therefore, this also closes #5.
Cleaned up the config folder.
Added SENDER to the list of environment variables needed and use that
for SMTP settings.
Cleaned up a few classes that were just noise.
Improved the help message